Return pointer from NewWriter.
Writer has a boolean member that indicates whether the first sentence
is written or not, in order to decide if a sentence separator should
be written. If the writer struct is passed by value, this information
gets lost, and no sentence separators are written.

To avoid accidentally running into this problem, we return a pointer
to the writer.